01-agentic-gateway/requirements.txt
vllm>=0.7.0
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
pyyaml>=6.0
pydantic>=2.10.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
